rename serializers to encoders (rename files)
