+++ silkworm/db/datastore/snapshots/common/codec.hpp
