+++ silkworm/db/transactions/txn_index.cpp
@@ -19,7 +19,7 @@
 #include <silkworm/db/blocks/bodies/body_txs_amount_query.hpp>
 #include <silkworm/db/datastore/snapshots/segment/segment_reader.hpp>
 
-#include "txn_segment_word_serializer.hpp"
+#include "txn_segment_word_codec.hpp"
 
 namespace silkworm::snapshots {
